fix: make hitBox match sprite size
The literal sprite sheet rendering config is store seperately from the
"hitbox" config for a given moving object. This resolves the issue where
the enemy was able clip through walls. This appeared to begin with
because on each frame the moving objects are moving little enough
distance that they can fit themselves in the gap between that mis-sized
hitbox and the actual sprite size.

Update MovingObjects.canMove() to make use of canvas coordinates in
method and letting map do the conversion. This should resolve clipping
for all movingobjects.

feat: implement deltaTime movement (secondsPassed)