fix: sprite offset glitching
Now using specific sprite sheet config for each enemy skin, walkCycle,
direction combination even if this results in repetition. This has
created a new problem in the form of squashing and stretching in between
the sprites that are wider/narrower than one another.

Updated sprite sheet to have consistent horizontal offsets between the
two groups of sprites (walk cycle one and two).

Minor style change in the form of using lib over Lib to be clear it is
not a class and is instead my pile of commonly used junk.